Ok obviously React components do not like to be tested.
