+++ lib/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import { render } from 'react-dom';
-
-
 import Application from './components/Application';
 
-// require('./styles/style');
+// require('./Scrabble')
+
 
 render(<Application/>, document.getElementById('application'));
