+++ lib/scoreWords.js
@@ -1,0 +1,19 @@
+
+
+const letterScores = {
+  A: 1, B: 3, C: 3, D: 2,
+  E: 1, F: 4, G: 2, H: 4,
+  I: 1, J: 8, K: 5, L: 1,
+  M: 3, N: 1, O: 1, P: 3,
+  Q: 10, R: 1, S: 1, T: 1,
+  U: 1, V: 4, W: 4, X: 8,
+  Y: 4, Z: 10
+};
+
+function scoreWords() {
+  // scoreWord() takes a word and returns the total score of the word.
+  // scoreWord('hello') should return 8.
+  // An empty string or any non-string (e.g. scoreWord('') or scoreWord(null)) should return 0.
+  // The function should strip any whitespace before or after the word. scoreWorld(' hello ') should be the same as scoreWord('hello').
+  // Score word should take an optional second argument for the word multiplier. scoreWord('hello', 2) should return 16.
+}
