+++ package.json
@@ -26,7 +26,7 @@
     "babel-polyfill": "^6.7.4",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
-    "babel-register": "^6.7.2",
+    "babel-register": "^6.14.0",
     "chai": "^3.5.0",
     "css-loader": "^0.23.1",
     "eslint": "^2.6.0",
