+++ test/index.js
@@ -1,8 +1,1 @@
-const assert = require('chai').assert;
-require('../lib/components/Application.js');
-
-describe('The test bundle', function() {
-  it('should be connected', function() {
-    assert(true);
-  });
-});
+require('./testScrabble');
